WaveSabreCore: simplify cos-table computation
We don't need to use run-time trig functions here, because the step
between each entry is the same. So instead, let's calculate the
coefficients of a 2D rotation matrix using constant expressions, which
should give the same result, barring any accumulation errors.

Since we're using double precision floats and the LUT is of limited
size, the max error accumulated are a negligible 2.22045e-14, compared
to the CRT cos() function.